Clang doesn't like mixing u_int with uint8_t, use the latter instead.

+++ src/btree/bt_delete.c
@@ -58,7 +58,7 @@ __wt_delete_page(WT_SESSION_IMPL *session, WT_REF *ref, bool *skipp)
 {
     WT_ADDR *ref_addr;
     WT_DECL_RET;
-    u_int previous_state;
+    uint8_t previous_state;
 
     *skipp = false;
 
@@ -157,7 +157,7 @@ __wt_delete_page_rollback(WT_SESSION_IMPL *session, WT_REF *ref)
 {
     WT_UPDATE **updp;
     uint64_t sleep_usecs, yield_count;
-    u_int current_state;
+    uint8_t current_state;
     bool locked;
 
     /* Lock the reference. We cannot access ref->page_del except when locked. */
